// storybook/stories/API/arg-types/LabelArgs.ts
export const LabelArgs: StorybookArgs = {
  angle: {
    description:
      'Text rotation angle in degrees.\nPositive values rotate clockwise, negative values rotate counterclockwise.',
    control: {
      type: 'number',
    },
    table: {
      type: {
        summary: 'number',
      },
      category: 'General',
      defaultValue: {
        summary: '0',
      },
    },
    defaultValue: 0,
  },
  children: {
    description: 'The value of label can be set as children or as the `value` prop',
    table: {
      type: {
        summary: 'ReactNode',
      },
      category: 'General',
    },
  },
  className: {
    control: {
      type: 'text',
    },
    table: {
      type: {
        summary: 'string',
      },
      category: 'Style',
    },
  },
  content: {
    description:
      'If set a React element, the option is the custom react element of rendering label.\nIf set a function, the function will be called to render label content.',
    table: {
      type: {
        summary: 'Function | ReactNode',
      },
      category: 'General',
    },
  },
  formatter: {
    description:
      'Function to customize how content is serialized before rendering.\n\nThis should return a renderable text - something that the {@link Text} component can render.\nTypically, a string or number.\nCustom components are not supported here - use the `content` prop instead.',
    table: {
      type: {
        summary: 'Function',
      },
      category: 'General',
    },
  },
  id: {
    description: 'Unique identifier of this component.\nUsed as an HTML attribute `id`.',
    control: {
      type: 'text',
    },
    table: {
      type: {
        summary: 'string',
      },
      category: 'General',
    },
  },
  index: {
    control: {
      type: 'number',
    },
    table: {
      type: {
        summary: 'number',
      },
      category: 'General',
    },
  },
  labelRef: {
    table: {
      type: {
        summary: 'React.RefObject<SVGTextElement> | null',
      },
      category: 'General',
    },
  },
  offset: {
    description: 'The offset to the specified "position". Direction of the offset depends on the position.',
    table: {
      type: {
        summary: 'number | string',
      },
      category: 'General',
      defaultValue: {
        summary: '5',
      },
    },
    defaultValue: 5,
  },
  parentViewBox: {
    table: {
      type: {
        summary: 'Required<CartesianViewBox> | Required<PolarViewBox>',
      },
      category: 'General',
    },
  },
  position: {
    description: 'The position of label relative to the view box.',
    table: {
      type: {
        summary:
          '"bottom" | "center" | "centerBottom" | "centerTop" | "end" | "inside" | "insideBottom" | "insideBottomLeft" | "insideBottomRight" | "insideEnd" | "insideLeft" | "insideRight" | "insideStart" | "insideTop" | "insideTopLeft" | "insideTopRight" | "left" | "middle" | "outside" | "right" | "top" | { x?: string | number | undefined; y?: string | number | undefined; }',
      },
      category: 'General',
      defaultValue: {
        summary: 'middle',
      },
    },
    defaultValue: 'middle',
  },
  textBreakAll: {
    control: {
      type: 'boolean',
    },
    table: {
      type: {
        summary: 'boolean',
      },
      category: 'General',
      defaultValue: {
        summary: 'false',
      },
    },
    defaultValue: false,
  },
  value: {
    description: 'The value of label can be set as children or as the `value` prop',
    table: {
      type: {
        summary: 'false | null | number | string | true',
      },
      category: 'General',
    },
  },
  viewBox: {
    description:
      'The box of viewing area. Used for positioning.\nIf undefined, viewBox will be calculated based on surrounding context.',
    table: {
      type: {
        summary: 'Required<CartesianViewBox> | Required<PolarViewBox>',
      },
      category: 'General',
    },
  },
  zIndex: {
    description:
      'Z-Index of this component and its children. The higher the value,\nthe more on top it will be rendered.\nComponents with higher zIndex will appear in front of components with lower zIndex.\nIf undefined or 0, the content is rendered in the default layer without portals.',
    control: {
      type: 'number',
    },
    table: {
      type: {
        summary: 'number',
      },
      category: 'General',
      defaultValue: {
        summary: '2000',
      },
    },
    defaultValue: 2000,
  },
};